10월/project_5/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
black==22.8.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
debugpy==1.6.0
decorator==5.1.1
Django==4.0.6
django-appconf==1.0.5
django-bootstrap5==22.1
django-extensions==3.2.0
django-imagekit==4.1.0
entrypoints==0.4
executing==0.10.0
h11==0.13.0
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.5
outcome==1.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pywin32==304
pyzmq==23.2.0
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.4.0
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.10
wcwidth==0.2.5
wsproto==1.2.0
